mini_wiki/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
pyyaml>=6.0
pydantic>=1.9.0
click>=8.0.0
rich>=10.0.0
textual>=0.1.0
python-dotenv>=0.19.0
PyPDF2>=1.26.0
requests>=2.27.0
beautifulsoup4>=4.10.0
lxml>=4.9.0
openpyxl>=3.7.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.18.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.950
pre-commit>=2.17.0
